configure script and Makefile workarounds
The superconfigure script calls configure with all the necessary
parameters for compilation.  Even then, the configure script would fail
checks for math functions (like asin, acos).  Reading through
configure.ac and the autoconf documentation showed the clash was with
AC_CHECK_FUNC -- the sample file used to check for such functions
declared all of them with the signature char func(); and this errored
with the included cosmopolitan.h which had the right signature for these
functions.

Running make clean was removing cosmopolitan.a as the clean target uses
a wildcard in find. Also changed the order of libraries to build the
python executable so that cosmopolitan would be last.

The libcosmo should folder contain the necessary cosmopolitan-related
files.