Static compilation of internal modules
setup.py only attempts to compile those modules that did not have a
preexisting compilation recipe. To compile these modules before setup.py
is called, one must change Modules/Setup, but that is ignored because it
is generated from Modules/Setup.dist. I changed Setup.dist to enable as
many modules as possible.

* the array module had a namespace clash (initarray), undef'd macro
* removed the zlib.h include becuase it is available in cosmopolitan
* mmap module errors with signature of remap
* syslog module errors (open/closelog, setlogmask, log_mask, log_upto)
* locale module compiles but linker error at strxfrm
* pyexpat module didn't compile because recipe was wrong
* _socket module needs netdb.h which isn't available yet

Also allowed Modules/Setup.local into the repo because that enables
statically compiling and linking (in theory) any third-party C library
in libpython.a if given a recipe with the right flags, files.