Python APE now runs scripts on Windows
Earlier the APE would open on Windows, but exit immediately, because it
was unable to find Lib/site.py which sets up the import locations for
the rest of the Python standard library.

The stdlib confusion with the APE occurs because Python has 3 levels of
OS-specific checks: build-time checks (via configure or some #defines),
run-time checks in compiled internal modules (like in getplatform.c or
getpath.c), and run-time checks in .py files (like os.py or site.py).

Since we are compiling on linux, the build-time OS is set to linux2,
which is what is returned by sys.platform (i.e. Py_GetPlatform).
Py_GetPlatform now uses the IsWindows/IsLinux checks from Cosmopolitan.

There was also an issue of hardcoded absolute paths in getpath.c, which
eventually provides the initial values for sys.path in Python, and
site.py. So those absolute paths were changed to relative paths. (I
expected issues with the path separators and forward slash/backslash,
but it didn't happen).

Finally, os.py decides what the underlying OS is based on what internal
modules are available in Python. Compiling on Linux means the posix
module is default, so I changed this check to use sys.platform instead.

Now the APE runs on Windows, but interactive mode throws a syntax error
for every line (setvbuf related?). Running scripts is fine though.