Avoid syntax error related to CRLF on Windows
The compiled executable displayed syntax errors at every line when run
on Windows in interactive mode. This was because CRLF was being fed to
Python when I pressed in the CMD console, and it was being parsed
incorrectly. I changed the tokenizer to ignore CR along with the
existing whitespace characters, and now no errors on Windows.