Embed python standard lib inside the APE
The python interpreter sets up the zipimport module at the very
beginning of execution, so that it can look for modules within any zip
files provided in sys.path. Cosmopolitan Libc allows using the APE
itself as a readable ZIP store.

I added locations within the internal ZIP store as the first entries of
sys.path, and used zip to embed the standard libraries + configdata into
the APE itself, and now python.com is self-contained.

It uses the relative location of the APE to figure out where to import
from, so this may not work if python.com is on $PATH.