Compile C extensions into the APE
Most packages for Python have internal binary extension modules (in C or
other languages). Since the APE does not have dlopen, and pip does not
work if SSL is not available, the only option is to download the source,
compile the binary extension alongside the APE itself, add the glue code
as a python-only module in Lib, and then change the
inits/imports/includes accordingly to ensure the extension has the
correct name everywhere.

I did this for the greenlet package: https://pypi.org/project/greenlet/,
downloaded the source, moved the C code to within Modules, added the
correct recipe to Modules/Setup, changed the headers, and changed the
imports in the glue code (both __init__.py and init_greenlet in greenlet.c)
to import the extension correctly.

It works. I am able to import the greenlet package in python.com without
error. Therefore one can embed C extension modules into the APE, but you
gotta do a lot of bookkeeping.