enable socketmodule on Windows
The simple call of python -m http.server was failing on Windows with a
ENOPROTOOPT error. Debugging showed that the issue was with getsockopt
passing a wrong option flag when on Windows: SO_REUSEADDR is defined as
0 for Windows in Cosmopolitan, and I think it should be 1.

The change here just checks if IsWindows and calls setsockopt with the
value of 1, and now http.server works on Windows as well.