Minor build system changes
I added -ffunction-sections, -fdata-sections, and --gc-sections to the
build, removed some tkinter-dependent packages from the stdlib (turtle
and IDLE), removed a couple of warnings, changed the link order so
cosmopolitan is linked last.

Now the APE without any libraries in the ZIP store is 4MB, APE with all
the necessary stdlib is 8.7MB (pyo files are not getting generated?).

I was able to compile OpenSSL from source, link libssl.a with python,
and use pip from APE to download packages (installing packages into the
ZIP causes ETXTBSY, because the APE cannot self-modify). Building
OpenSSL from source with Cosmopolitan is left as an exercise to the
reader.