added _decimal module and libmpdec to build
I didn't see that libmpdec was included in the _decimal module, could've
added it quite a while back. The _decimal module adds some KBs to the
APE but the tradeoff is that I can ignore _pydecimal.py now, so it
balances out.

Minimal APE size: without pip, setuptools, asyncio, and unicodedata,
the size of the APE is 5MB. Can possibly remove distutils + msilib
maybe, to reduce size further.